requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
django-allauth==0.45.0
django-celery-results==2.1.0
django-cors-headers==3.10.0
django-grappelli==2.15.1
django-oauth-toolkit==1.5.0
django-redis==5.0.0
django-rest-knox==4.1.0
Django==3.2.8
djangorestframework==3.12.4
idna==3.3
oauthlib==3.1.1
Pillow==8.4.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.3.0
python-http-client==3.3.3
python3-openid==3.2.0
pytz==2021.3
requests-oauthlib==1.3.0
requests==2.26.0
sendgrid==6.8.2
sqlparse==0.4.2
starkbank-ecdsa==2.0.0
twilio==7.1.0
urllib3==1.26.7
